refactor: extract useTriggersState hook from AppContext
Move 6 trigger useState calls + 8 trigger callbacks (loadTriggers,
loadTriggerHealth, loadTriggerRuns, createTrigger, updateTrigger,
deleteTrigger, runTriggerNow, sortTriggersByNextRun) into a standalone
useTriggersState hook. Follows the existing useLifecycleState pattern.

AppContext.tsx: 8,366 → 8,199 lines (-167)
useTriggersState.ts: 215 lines (new)

Zero coupling to startup effect — triggers are only loaded post-ready.
Context value shape unchanged — consumers unaffected.
